perf/refactor: added burst compile attrib to all blittable type utility methods, add MeshJobCore unified, etc...

+++ Assets/Scripts/Chunk/ChunkData.cs
@@ -7,10 +7,10 @@ namespace Chunk
     [StructLayout(LayoutKind.Sequential)]
     public struct ChunkData : IEquatable<ChunkData>
     {
-        public float x;
-        public float z;
+        public int x;
+        public int z;
 
-        public ChunkData(float x, float z)
+        public ChunkData(int x, int z)
         {
             this.x = x;
             this.z = z;
